refactor: Change data structures
Before the images, features and labels where saved as 'vehicles' and 'non-vehicles'. This caused difficulties when shuffling and splitting the dataset into 'train' and 'test' sets, because it was complicated to relate the shuffled features to the images.
Now the files of vehicles and non-vehicles are concatenated and shuffled, before generating features. This way it is always possible to tell which image belongs to which feature.